app/requirements.txt
Django==4.2.28
djangorestframework==3.16.1
gunicorn==25.0.3
django-environ==0.12.0
boto3==1.42.44
whitenoise==6.6.0
